Делит функции по модулям
