+++ index.html
@@ -228,8 +228,7 @@
   </template>
 
   <!-- Подключение скриптов -->
-  <script src="js/main.js"></script>
-  <script src="js/functions.js"></script>
+  <script src="js/main.js" type="module"></script>
 
 </body>
 </html>
